Add: gitignore file to ignore binary files
The gitignore file will automatically ignore all the binary files generated by the visual studio when the program the executed. The binary files are generated when the program is compiled and are not required to be committed. Instead, they will cause many conflicts when multiple people will start collaborating on this repository.